backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
flask-bcrypt==1.0.1
python-dotenv==1.0.1
gunicorn==21.2.0
